Add files via upload
An analysis of the most recent data for UK divorce rates for both men and women. The analysis covers various topics of comparisons (between men and women) related to divorce, such as divorce submissions, the mean age of marriage, etc. 

The data was sourced from kaggle, however, the original source of the data is from the UK Office of National  Statistics (ONS).